Create Swap Service
