+++ Want2LearnBase/Program.cs
@@ -23,7 +23,7 @@ namespace Want2LearnBase
     {
         static void Main(string[] args)
         {
-            
+           
         }
 
         public static int[] RandomArray(int number, Random rnd)
@@ -68,53 +68,6 @@ namespace Want2LearnBase
            
         }
 
-        public static void Swap(ref string line1, ref string line2)
-        {
-            string Temp;
-            Temp = line1;
-            line1 = line2;
-            line2 = Temp;
-        }
-
-        public static void Swap(ref int first, ref int second)
-        {
-            int Temp;
-            Temp = first;
-            first = second;
-            second = Temp;
-        }
-
-        public static void Swap(ref double first, ref double second)
-        {
-            double Temp;
-            Temp = first;
-            first = second;
-            second = Temp;
-        }
-
-        public static void Swap(ref decimal first, ref decimal second)
-        {
-            decimal Temp;
-            Temp = first;
-            first = second;
-            second = Temp;
-        }
-
-        public static void Swap(ref byte first, ref byte second)
-        {
-            byte Temp;
-            Temp = first;
-            first = second;
-            second = Temp;
-        }
-
-        public static void Swap(ref float first, ref float second)
-        {
-            float Temp;
-            Temp = first;
-            first = second;
-            second = Temp;
-        }
 
         public static string GetAddress(string city, string street, string houseNumber, string flatNumber)
         {
