+++ Want2LearnBase/SwapService.cs
@@ -1,0 +1,59 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Want2LearnBase
+{
+    public static class SwapService
+    {
+        public static void Swap(ref string line1, ref string line2)
+        {
+            string Temp;
+            Temp = line1;
+            line1 = line2;
+            line2 = Temp;
+        }
+
+        public static void Swap(ref int first, ref int second)
+        {
+            int Temp;
+            Temp = first;
+            first = second;
+            second = Temp;
+        }
+
+        public static void Swap(ref double first, ref double second)
+        {
+            double Temp;
+            Temp = first;
+            first = second;
+            second = Temp;
+        }
+
+        public static void Swap(ref decimal first, ref decimal second)
+        {
+            decimal Temp;
+            Temp = first;
+            first = second;
+            second = Temp;
+        }
+
+        public static void Swap(ref byte first, ref byte second)
+        {
+            byte Temp;
+            Temp = first;
+            first = second;
+            second = Temp;
+        }
+
+        public static void Swap(ref float first, ref float second)
+        {
+            float Temp;
+            Temp = first;
+            first = second;
+            second = Temp;
+        }
+    }
+}
