+++ Want2LearnBase/Want2LearnBase.csproj
@@ -45,6 +45,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SwapService.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
